Move game player to separate function
When making changes to the search algorithm, we need to have
accurate ways to determine if the changes actually improved
performance. One way to determine this is by playing multiple
games and seeing if the resulting score distributions are
different. To further this end, the game playing function has
been moved to a separate function.