Add in probability based pruning
Prune the search when the probability of reaching a state is below
a human defined threshold. This focuses the search on states which
are more likely to occur, and terminates subtrees early when they
arise from improbable events (e.g. multiple 4s spawning).

This allows us to have a higher upper bound on the maximum search
depth for more probable events.