Add search time tracker and adjust parameters
Our previous parameters of looking up to 8 moves ahead causes the
program to be unbearably slow. It can take up to a full day to play
a single game, which is unacceptable if we want to test whether the
heuristics are making a real difference (this requires multiple
runs).

The maximum search depth was scaled back and a timer was added to
check how long the search function is taking for every move.